[mod]refactor
Separate std wrapper from errors.go
