+++ errors.go
@@ -1,7 +1,6 @@
 package errors
 
 import (
-	"errors"
 	"fmt"
 	"sync"
 )
@@ -56,11 +55,3 @@ func (e *fundamental) Format(f fmt.State, c rune) {
 		fmt.Fprint(f, e.Error())
 	}
 }
-
-// Is the wrapper for the standard errors.Is().
-// It's returns same result.
-func Is(err, target error) bool { return errors.Is(err, target) }
-
-// As the wrapper for the standard errors.As().
-// It's returns same result.
-func As(err error, target any) bool { return errors.As(err, target) }
